custom_components/glados_tts_remote/pyproject.toml
[tool.poetry]
name = "glados-tts-remote"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiodns = "3.2.0"
aiohasupervisor = "0.2.1"
aiohttp = "3.11.8"
aiohttp_cors = "0.7.0"
aiohttp-fast-zlib = "0.2.0"
aiozoneinfo = "0.2.1"
astral = "2.2"
async-interrupt = "1.2.0"
attrs = "24.2.0"
atomicwrites-homeassistant = "1.4.1"
audioop-lts = { version = "0.2.1", python = "^3.13" }
awesomeversion = "24.6.0"
bcrypt = "4.2.0"
certifi = "^2021.5.30"
ciso8601 = "2.3.1"
fnv-hash-fast = "1.0.2"
hass-nabucasa = "0.85.0"
httpx = "0.27.2"
home-assistant-bluetooth = { version = "1.13.0", python = ">=3.12,<3.14"}
ifaddr = "0.2.0"
Jinja2 = "3.1.4"
lru-dict = "1.3.0"
PyJWT = "2.10.0"
cryptography = "43.0.1"
Pillow = "11.0.0"
propcache = "0.2.0"
pyOpenSSL = "24.2.1"
orjson = "3.10.12"
packaging = ">23.1"
psutil-home-assistant = "0.0.1"
python-slugify = "8.0.4"
PyYAML = "6.0.2"
requests = "2.32.3"
securetar = "2024.11.0"
SQLAlchemy = "2.0.31"
standard-aifc = { version = "3.13.0", python = "^3.13" }
standard-telnetlib = { version = "3.13.0", python = "^3.13" }
typing-extensions = ">=4.12.2,<5.0"
ulid-transform = "1.0.2"
urllib3 = ">1.26.5,<2"
uv = "0.5.4"
voluptuous = "0.15.2"
voluptuous-serialize = "2.6.0"
voluptuous-openapi = "0.0.5"
yarl = "1.18.0"
webrtc-models = "0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
